% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lineups}
\alias{get_lineups}
\title{Read CSV and output players and stacks}
\usage{
get_lineups(dkDataset, playerSalaryDataset, salaryType)
}
\arguments{
\item{dkDataset}{Dataset of Draftkings tournament result}
}
\value{
A dataframe with players and stacks used
}
\description{
Read CSV and output players and stacks
}
\examples{
\dontrun{

}

}
