% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_lineup_positions}
\alias{parse_lineup_positions}
\title{Parse through dataframe to extract player names from lineups}
\usage{
parse_lineup_positions(entryDf)
}
\arguments{
\item{entryDf}{Draftkings Dataframe of entry information containing lineups}
}
\value{
none
}
\description{
Parse through dataframe to extract player names from lineups
}
\examples{

\dontrun{
dk_parsed = parse_lineup_positions(entry_df)
}

}
