% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_with_index}
\alias{replace_with_index}
\title{Add indexing to positions where there are multiple, i.e. OF and P}
\usage{
replace_with_index(string, tag)
}
\arguments{
\item{string}{Lineup cell to be extracted}

\item{tag}{Specific position to index}
}
\value{
A dataframe with indexed positions added to the Lineup
}
\description{
Add indexing to positions where there are multiple, i.e. OF and P
}
\examples{
\dontrun{
mutate(Lineup = sapply(Lineup, function(l) {
 l \%>\%
  replace_with_index("OF") \%>\%
  replace_with_index("P")
}))
}


}
