% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_team_stacks}
\alias{summarize_team_stacks}
\title{Summarize team stacks}
\usage{
summarize_team_stacks(dkCombined)
}
\arguments{
\item{dkCombined}{DK Dataframe of all players and info added}
}
\value{
none
}
\description{
Summarize team stacks
}
\examples{
\dontrun{
summarize_team_stacks(wide_lineup)
}
}
