% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{widen_lineup}
\alias{widen_lineup}
\title{Widen lineup to add fantasy points, salary and team}
\usage{
widen_lineup(longLineupDf)
}
\arguments{
\item{longLineupDf}{Dataframe of lineup with player information in a long format}
}
\value{
none
}
\description{
Widen lineup to add fantasy points, salary and team
}
\examples{
\dontrun{
wide_lineup = widen_lineup(longLineupDf)
}
}
